Move Many Types to types.ts - move all types from typegoose.ts to types.ts - move many types from prop.ts to types.ts - add tsdoc for some types - typegoose.ts: rename t argument to cl - set "removeComments" rule to false to add tsdoc comments to .d.ts
Note:
keep "removeComments" as "false" as long as microsoft/TypeScript#14619 isnt fixed
