+++ tsconfig.json
@@ -4,7 +4,7 @@
     "lib": [
       "esnext"
     ],
-    "removeComments": true,
+    "removeComments": false,
     "module": "commonjs",
     "moduleResolution": "node",
     "newLine": "LF",
